[#33341] docdb: Prefer shared memory whenever possible for lock acquire and release
Summary:
In order to avoid postgres <=> TServer context switch on object lock acquisition, there is a
fastpath for object locking that simply writes the lock request into a buffer in shared memory
(`ObjectLockSharedState`) and returns immediately on lock acquition (with appropriate restrictions
to ensure this is safe). On all TServer-side operations for object locking, we consume the entire
buffer of lock requests into `ObjectLockManager` then proceed with the operation.

This still has a significant overhead for object locks in some cases:
- high session count results in mutex contention on the global mutex in `ObjectLockManager`
- read-only transactions still need to an additional postgres <=> TServer roundtrip for
  FinishTransaction that was not necessary before object locking

To address this overhead, we have moved to one `ObjectLockSharedState` buffer per postgres session
with d195c5c0fdece04bbc1e1d744337566544136342. This revision applies the following optimizations
that are now possible:
- Read-only transactions now skip `FinishTransaction` (like in the table locks off case) whenever
  possible.
  - Specifically, as long as there is no entry for the transaction in `ObjectLockManager`
    (created via acquire lock RPC or when consuming lock requests for a possibly unrelated
    `ObjectLockManager` call), it is now possible to release all locks by simply dropping the lock
    request buffer for the session.
- If `FinishTransaction` must be sent because there was a write, we attempt to release locks without
  going to `ObjectLockManager` whenever possible.
  - Same conditions as above case.
- To avoid creating an entry in `ObjectLockManager` in the case where the plain transaction has not
  yet been created at lock acquire time (and thus acquire lock RPC is required), TServer attempts
  to acquire shared locks by writing to the session-specific shared memory buffer first before
  going to `ObjectLockManager`.
- To avoid lock releases via `ObjectLockManager` cascading into other `ObjectLockManager` lock
  releases for unrelated transaction, we now only consume lock requests for the transaction
  releasing locks. In the case where we are releasing all locks for the transaction, this has been
  changed to just drop all locks for the transaction.

Other related changes:
- OID fields for locking related protobufs were changed to uint64 to uint32, to avoid unnecessary
  casting. OIDs are 32-bit, so there was no need for these fields to be uint64.
- The `object_locking_fastpath_acquires` metric was renamed to `object_locking_fastpath_pg_acquires`
  to distinguish it from the the newly added `object_locking_fastpath_tserver_acquires`. These
  measure the number of lock acquires done via shared memory, from postgres and TServer side
  respectively.
- `object_locking_fastpath_pg_releases`, `object_locking_fastpath_tserver_releases`, and
  `object_locking_lock_releases` were added as a counterpart for the existing acquire metrics.

**Upgrade/Rollback safety**

OID fields for locking related protobufs were changed from uint64 to uint32. This is binary
compatible, and OIDs were always 32-bit anyways.

Test Plan:
Added tests:
```
./yb_build.sh --cxx-test ts_local_lock_manager-test --gtest_filter 'TSLocalLockManagerTest.TestFastpathRelease*'
./yb_build.sh --cxx-test pg_object_locks-test --gtest_filter 'PgObjectLocksFastpathTest.*'
```

The `PgObjectLocksFastpathTest.TestSimple` was removed since it no longer applies.

The following performance tests were also run.
```
 Workload                      | master (base) | master (locks) |    diff (base) |   diff (locks) |
  oltp_read_only
   - Latency (ms)              |         12.37 |  12.98 (+4.9%) |  12.23 (-1.1%) |  12.24 (-1.1%) |
   - Throughput (events/s)     |          4848 |   4622 (-4.7%) |   4907 (+1.2%) |   4901 (+1.1%) |
  oltp_read_write
   - Latency (ms)              |         84.34 |  86.72 (+2.8%) |  84.29 (-0.1%) |  84.81 (+0.6%) |
   - Throughput (events/s)     |        284.54 | 276.75 (-2.7%) | 284.72 (+0.1%) | 282.99 (-0.5%) |
  THRPT_multivalue_insert_1row
   - Latency (ms)              |          1.41 |  1.68 (+19.1%) |   1.41 (+0.0%) |   1.36 (-3.5%) |
   - Throughput (requests/s)   |         25432 | 21406 (-15.8%) |  25457 (+0.0%) |  26402 (+3.8%) |
```

Reviewers: bkolagani

Reviewed By: bkolagani

Subscribers: yql, ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D56889

+++ src/yb/integration-tests/object_lock-test.cc
@@ -278,20 +278,20 @@ auto kTxn1 = docdb::ObjectLockOwner{TransactionId::GenerateRandom(), 1};
 auto kTxn2 = docdb::ObjectLockOwner{TransactionId::GenerateRandom(), 1};
 auto kTxn3 = docdb::ObjectLockOwner{TransactionId::GenerateRandom(), 1};
 
-constexpr uint64_t kDatabaseID = 1;
-constexpr uint64_t kRelationId = 1;
-constexpr uint64_t kRelationId2 = 2;
-constexpr uint64_t kRelationId3 = 3;
-constexpr uint64_t kDefaultObjectId = 0;
-constexpr uint64_t kDefaultObjectSubId = 0;
+constexpr uint32_t kDatabaseID = 1;
+constexpr uint32_t kRelationId = 1;
+constexpr uint32_t kRelationId2 = 2;
+constexpr uint32_t kRelationId3 = 3;
+constexpr uint32_t kDefaultObjectId = 0;
+constexpr uint32_t kDefaultObjectSubId = 0;
 constexpr uint64_t kLeaseEpoch = 1;
 const MonoDelta kTimeout = MonoDelta::FromSeconds(8);
 constexpr auto kDefaultTestStatusTabletId = "test_status_tablet";
 
 template <typename Request>
 Request AcquireRequestFor(
-    const std::string& session_host_uuid, const docdb::ObjectLockOwner& owner, uint64_t database_id,
-    uint64_t relation_id, TableLockType lock_type, uint64_t lease_epoch, server::ClockPtr clock,
+    const std::string& session_host_uuid, const docdb::ObjectLockOwner& owner, uint32_t database_id,
+    uint32_t relation_id, TableLockType lock_type, uint64_t lease_epoch, server::ClockPtr clock,
     std::optional<HybridTime> deadline, TransactionId background_txn) {
   Request req;
   owner.PopulateLockRequest(&req);
@@ -351,7 +351,7 @@ void CallProxyMethod(
 template <typename Request, typename Response, typename Proxy>
 std::future<Status> AcquireLockAsync(
     Proxy* proxy, const std::string& session_host_uuid, const docdb::ObjectLockOwner& owner,
-    uint64_t database_id, uint64_t relation_id, TableLockType type, uint64_t lease_epoch,
+    uint32_t database_id, uint32_t relation_id, TableLockType type, uint64_t lease_epoch,
     server::ClockPtr clock, std::optional<HybridTime> opt_deadline, MonoDelta rpc_timeout,
     TransactionId background_txn) {
   auto resp = std::make_shared<Response>();
@@ -380,7 +380,7 @@ std::future<Status> AcquireLockAsync(
 
 std::future<Status> AcquireLockAsyncAt(
     tserver::TabletServerServiceProxy* proxy, const std::string& session_host_uuid,
-    const docdb::ObjectLockOwner& owner, uint64_t database_id, uint64_t relation_id,
+    const docdb::ObjectLockOwner& owner, uint32_t database_id, uint32_t relation_id,
     uint64_t lease_epoch = kLeaseEpoch, server::ClockPtr clock = nullptr,
     std::optional<HybridTime> deadline = std::nullopt, MonoDelta rpc_timeout = kTimeout,
     TransactionId background_txn = TransactionId::Nil()) {
@@ -393,7 +393,7 @@ std::future<Status> AcquireLockAsyncAt(
 
 Status AcquireLockAt(
     tserver::TabletServerServiceProxy* proxy, const std::string& session_host_uuid,
-    const docdb::ObjectLockOwner& owner, uint64_t database_id, uint64_t relation_id,
+    const docdb::ObjectLockOwner& owner, uint32_t database_id, uint32_t relation_id,
     uint64_t lease_epoch = kLeaseEpoch, server::ClockPtr clock = nullptr,
     std::optional<HybridTime> deadline = std::nullopt, MonoDelta rpc_timeout = kTimeout,
     TransactionId background_txn = TransactionId::Nil()) {
@@ -405,7 +405,7 @@ Status AcquireLockAt(
 
 std::future<Status> AcquireLockGloballyAsync(
     master::MasterDdlProxy* proxy, const std::string& session_host_uuid,
-    const docdb::ObjectLockOwner& owner, uint64_t database_id, uint64_t relation_id,
+    const docdb::ObjectLockOwner& owner, uint32_t database_id, uint32_t relation_id,
     uint64_t lease_epoch = kLeaseEpoch, server::ClockPtr clock = nullptr,
     std::optional<HybridTime> deadline = std::nullopt, MonoDelta rpc_timeout = kTimeout,
     TransactionId background_txn = TransactionId::Nil()) {
@@ -418,7 +418,7 @@ std::future<Status> AcquireLockGloballyAsync(
 
 Status AcquireLockGlobally(
     master::MasterDdlProxy* proxy, const std::string& session_host_uuid,
-    const docdb::ObjectLockOwner& owner, uint64_t database_id, uint64_t relation_id,
+    const docdb::ObjectLockOwner& owner, uint32_t database_id, uint32_t relation_id,
     uint64_t lease_epoch = kLeaseEpoch, server::ClockPtr clock = nullptr,
     std::optional<HybridTime> deadline = std::nullopt, MonoDelta rpc_timeout = kTimeout,
     TransactionId background_txn = TransactionId::Nil()) {
@@ -430,7 +430,7 @@ Status AcquireLockGlobally(
 
 std::future<Status> AcquireLockGloballyAsync(
     client::YBClient* client, const std::string& session_host_uuid,
-    const docdb::ObjectLockOwner& owner, uint64_t database_id, uint64_t relation_id,
+    const docdb::ObjectLockOwner& owner, uint32_t database_id, uint32_t relation_id,
     uint64_t lease_epoch = kLeaseEpoch, std::optional<HybridTime> opt_deadline = std::nullopt,
     MonoDelta rpc_timeout = kTimeout, TransactionId background_txn = TransactionId::Nil()) {
   auto promise = std::make_shared<std::promise<Status>>();
@@ -529,7 +529,7 @@ Status ReleaseLockGlobally(
 
 std::future<Status> WaitForLockersGloballyAsync(
     master::MasterDdlProxy* proxy, const std::string& session_host_uuid,
-    uint64_t database_id, uint64_t relation_id, TableLockType lock_type,
+    uint32_t database_id, uint32_t relation_id, TableLockType lock_type,
     uint64_t lease_epoch = kLeaseEpoch, MonoDelta rpc_timeout = kTimeout) {
   auto resp = std::make_shared<master::WaitForLockersMultipleGlobalResponsePB>();
   auto controller = std::make_shared<rpc::RpcController>();
@@ -560,7 +560,7 @@ std::future<Status> WaitForLockersGloballyAsync(
 
 std::future<Status> WaitForLockersGloballyAsync(
     client::YBClient* client, const std::string& session_host_uuid,
-    uint64_t database_id, uint64_t relation_id, TableLockType lock_type,
+    uint32_t database_id, uint32_t relation_id, TableLockType lock_type,
     uint64_t lease_epoch = kLeaseEpoch, MonoDelta rpc_timeout = kTimeout) {
   auto promise = std::make_shared<std::promise<Status>>();
   auto future = promise->get_future();
@@ -811,14 +811,14 @@ TEST_F(ObjectLockTest, DDLLockWaitsAtMaster) {
 
 TEST_F(ObjectLockTest, DDLLocksCleanupAtMaster) {
   auto master_proxy = ASSERT_RESULT(MasterLeaderProxy());
-  constexpr uint64_t kNumHosts = 2;
-  constexpr uint64_t kNumDDLsPerHost = 3;
-  constexpr uint64_t kNumObjectsPerDDL = 3;
-  constexpr uint64_t kLocksPerHost = kNumDDLsPerHost * kNumObjectsPerDDL;
-  constexpr uint64_t kNumLocksTotal = kLocksPerHost * kNumHosts;
+  constexpr uint32_t kNumHosts = 2;
+  constexpr uint32_t kNumDDLsPerHost = 3;
+  constexpr uint32_t kNumObjectsPerDDL = 3;
+  constexpr uint32_t kLocksPerHost = kNumDDLsPerHost * kNumObjectsPerDDL;
+  constexpr uint32_t kNumLocksTotal = kLocksPerHost * kNumHosts;
 
   auto ddl_txns = CreateRandomExclusiveLockOwners(kNumDDLsPerHost * kNumHosts);
-  for (uint64_t relation_id = 0; relation_id < kNumLocksTotal; relation_id++) {
+  for (uint32_t relation_id = 0; relation_id < kNumLocksTotal; relation_id++) {
     auto host_idx = relation_id / kLocksPerHost;
     auto ddl_idx = (relation_id / kNumObjectsPerDDL) % kNumDDLsPerHost;
     ASSERT_OK(AcquireLockGlobally(
